Large refactor.
Moved data structures into the structs.py module and defined some
supporting data structures.

Moved validation into it's own module.

Line now only accepts a string interface for it's data.

Panel now uses a tuple to store it's lines and reconstructs
it's collection of lines when it's rect is changed.